feat(tensor): implement TensorBuilder and TensorBuildError for tensor creation
