+++ api/tensor/src/lib.rs
@@ -1,67 +1,10 @@
-use dense_impl::{DenseBuildError, DenseBuilder, DenseTensorImpl};
+mod builder;
+mod tag;
+mod tensor;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TensorTag {
-    Dense,
-}
-
-#[derive(Debug, Clone)]
-pub enum Tensor {
-    Dense(DenseTensorImpl),
-}
-
-impl Tensor {
-    pub fn tag(&self) -> TensorTag {
-        match self {
-            Self::Dense(_) => TensorTag::Dense,
-        }
-    }
-
-    pub fn shape(&self) -> &[usize] {
-        match self {
-            Self::Dense(t) => t.shape(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TensorBuilder {
-    Dense(DenseBuilder),
-}
-
-impl TensorBuilder {
-    pub fn dense(shape: Vec<usize>) -> Self {
-        Self::Dense(DenseBuilder::new(shape))
-    }
-
-    pub fn tag(&self) -> TensorTag {
-        match self {
-            Self::Dense(_) => TensorTag::Dense,
-        }
-    }
-
-    pub fn build(self) -> Result<Tensor, TensorBuildError> {
-        match self {
-            Self::Dense(builder) => builder
-                .build()
-                .map(Tensor::Dense)
-                .map_err(TensorBuildError::Dense),
-        }
-    }
-}
-
-impl TensorTag {
-    pub fn builder(self, shape: Vec<usize>) -> TensorBuilder {
-        match self {
-            Self::Dense => TensorBuilder::dense(shape),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TensorBuildError {
-    Dense(DenseBuildError),
-}
+pub use builder::{TensorBuilder, TensorBuildError};
+pub use tag::TensorTag;
+pub use tensor::Tensor;
 
 #[cfg(test)]
 mod tests {
@@ -85,7 +28,6 @@ mod tests {
         match tensor {
             Tensor::Dense(inner) => {
                 assert_eq!(inner.shape(), &[1, 3]);
-                assert_eq!(inner.data(), vec![0.0, 0.0, 0.0]);
             }
         }
     }
